Add ChallengeRDG, adjust query String formatting
Create a `COLUMNS` string in the RDGs because I was using it often in queries.
Added the create table method for the `challenges` table in the Main class.
Deleted UserServlet because we won't be needing it. I was just using it as practice.
Set the `autocomplete` attribute in the register/login forms to `off`.

Signed-off-by: vartanbeno <vartanbeno@gmail.com>