Add ListPlayers servlet
The ListPlayers servlet returns, in JSON format, a list of users from the database. We just see the id and username of the user.
Added the JSTL jar file so that I can use the JSTL core tag. Useful for iterating through lists and making if statements in JSPs.
Changed the coontentType of the success/fail JSPs to application/json, so that the browser parses it as JSON.

Signed-off-by: vartanbeno <vartanbeno@gmail.com>