Add ChallengePlayer servlet
The ChallengePlayer servlet is used to challenge players.
You have to be logged in to access it. If you try to reach this page without logging in first, your will be redirected to /Login.
We check if you're logged in by checking if the `userid` attribute is available in the session.

There is a dropdown with a list of users that you pick from. Since you can't challenge yourself, your username does not appear.
Once you've thrown a challenge, a success message should appear.

Signed-off-by: vartanbeno <vartanbeno@gmail.com>