Create PageController parent servlet
Its init() method is used to set up the connection to the database, instead of having to do it in every servlet.
It also contains a method that closes the DB connection.
Also made methods to easily forward the user to the success/fail JSPs with an appropriate message.
Some more error handling, i.e. if you try to log out without loggin in first, you'll get a fail message.

Signed-off-by: vartanbeno <vartanbeno@gmail.com>