Passing in UserHelper object to JSPs
UserRDG instances, which belong in the technical services layer (bottom), shouldn't directly communicate with JSPs (template views), which belong in the presentation layer (top).
Instead, using the UserHelper, which belongs in the domain layer (middle), we eliminate the dependencies from the presentation layer to the technical services layer (top to bottom).
Also fixed a typo.

Signed-off-by: vartanbeno <vartanbeno@gmail.com>