Create AcceptChallenge servlet
Navigating to /AcceptChallenge provides a list of open challenges to you that you can accept.
Created the JSP for this page, called accept-challenge-form.jsp.
Created a method in ChallengeRDG to find an open challenge by the challenger's ID, and the challengee's ID.

Once you accept a challenge, you get a success message.
You can also get a failure message if you happen to send invalid data, such as an invalid combination of IDs, or if you accept a challenge against yourself.

Created the ChallengeHelper object, to separate the interaction between the presentation and the data layers (top <-> bottom).

Signed-off-by: vartanbeno <vartanbeno@gmail.com>