GetDeck servlet lets you create the deck
If you don't have a deck, you'll get a form which lets you create a deck (40 cards at random).
If you do have a deck, you'll get it in JSON format.
If you try to access /GetDeck and you're not logged in, you get a fail message.

Fixed a logic error in the CardsInDeckRDG findByDeck() method.

Signed-off-by: vartanbeno <vartanbeno@gmail.com>