Delete CardRDG, CardType, modify CardsInDeckRDG
We should not have a table of cards.

To upload a deck, the user him/herself passes in a list of cards.
For each, they decide the type (e, p, t) and the name.

To reflect this, the cards_in_deck table, instead of pointing to a card ID, instead has type/name attributes.
This is because cards aren't their own entity.

Renamed /GetDeck to /UploadDeck.

Signed-off-by: vartanbeno <vartanbeno@gmail.com>