Make username max length 64 instead of 30
One of the tests in the test suite provided had a registration with a really long username.
It was failing because it was never registering the user in the db, so I allow longer usernames now.

Created a Reset class, whose main method runs Teardown and Setup. Good for starting over in one run.

Renamed the 'challengee' parameter to 'player' when you challenge a player, to adhere to the test suite's names.

Signed-off-by: vartanbeno <vartanbeno@gmail.com>