Fix logic error in /ViewDeck
Was trying to find the user by the deck ID, and not the user ID...

Also, as it is now, if you make a GET request to /ChallengePlayer and don't have a deck, you get a fail message.
Made it so that you get the same result if you make a POST request to it.

Signed-off-by: vartanbeno <vartanbeno@gmail.com>