Add ViewBoard serlvet and view-board JSP
Go to /ViewBoard?game=[id] to see the board for the game with the specified ID.

If you don't specify a game ID, you will get a fail message.
The ID must be a number. If the format is off, you will get a fail message.
If a game with the specified ID doesn't exist, you will get a fail message.

You can only see the board if you are logged in.
You can only see the board if you are part of the game.

The board still doesn't contain all the necessary information. Just ID's of the players/decks.
Will add the rest later.

Also changed the variable for the view-deck.jsp from DECK -> VIEW_DECK.

Signed-off-by: vartanbeno <vartanbeno@gmail.com>