Error handling for DrawCard
If you try to draw a card in an invalid game, you get a fail message.
If you try to draw a card in a game you're not a part of, you get a fail message.
If you try to draw a card but you have no more cards left in your deck, you get a fail message.

Once you successfully draw a card, you get a success message, with the card you drew.
Goes without saying that you need to be logged in for all this to work.

Signed-off-by: vartanbeno <vartanbeno@gmail.com>