Create CardInPlayRDG for hand, benched & discarded
Deleted HandCardRDG and the hands table.

Instead of making a table for each of hands, benched, and discarded, I decided to make one table, called cards_in_play.
This table includes cards that have been drawn from the deck during a game.

It has a status attribute, which is an int with the following possible values:
	- 0: it's in the player's hand.
	- 1: it's in the player's bench. Only cards of type 'p' (Pokemon) can go on the bench. Have not implemented that logic yet.
	- 2: it's in the discard pile.
These are all the ordinal() values from the CardStatus enum I've created.

This big change came with a lot of code refactoring.

Fixes #3.

Signed-off-by: vartanbeno <vartanbeno@gmail.com>