Create PlayPokemonToBench servlet
Making a GET request to this yields an HTML form, with Pokemon (type 'p') cards that are in your hand.
You must specify the 'game' attribute, which is just the game ID.
You will get a fail message if:
	- you are not logged in.
	- you are not part of the game.
	- you didn't provide a valid format for the game ID (positive integer).
	- your bench already has 5 Pokemon on it.

Making a POST request (once you choose the card) will yield either a success or fail message.
You must specify the 'game' attribute like above, as well as a 'card' attribute, which is the ID of the card in the cards table that you want to pass on to the bench.
You will get a fail message for any of the reasons above, as well as if:
	- the card you're sending to the bench isn't of type 'p'.
	- the card you're sending to the bench is already on the bench.
	- the card you're sending to the bench is not in your hand.

Also made some other small changes not worth mentioning.
About to implement discarding cards.

Signed-off-by: vartanbeno <vartanbeno@gmail.com>