Create Retire servlet
Going to /Retire?game=[id] with a valid game ID that you're a part of will retire you from the game.
Once you've retired, you lose the game and the game is over.
You cannot draw any cards or play Pokemon to the bench in the game anymore. But you can still view its board.

Created a 'status' attribute for games. It can be any of the following values:
	- 0: the game is ongoing.
	- 1: the challenger retired.
	- 2: the challengee retired.
	- 3: the challenger won.
	- 4: the challengee won.
These are used in conjunction with the GameStatus enum.

Also updated README.md.

Signed-off-by: vartanbeno <vartanbeno@gmail.com>