Use card position as parameter for PlayPokemonToBench
Before, we were using the card ID as a parameter for PlayPokemonToBench. We were first checking if the card is in the player's hand.
Remember, it's /PlayPokemonToBench?game=[id]&card[position]

Now, instead of passing the card ID as the parameter, we will pass the position of the card in the player's hand.
We use a zero-based index. If your hand is, for example, [2, 5, 9, 12, 27] (these are the actual card IDs):
    - If you choose card 0, you get the first card, 2.
    - If you choose card 2, you get the third card, 9.
    - If you choose card 10, you should get a fail message.
Of course, if you don't have any cards in your hand to begin with, you get a fail message.

Also updated README.md.

Signed-off-by: vartanbeno <vartanbeno@gmail.com>