Fix error with closing statements
When trying to fetch domain objects with the Mappers, there was an error. Something like 'ResultSet is closed'.
That's because in the TDG methods. I was closing the PreparedStatement, THEN return the ResultSet.
But when you close a statement, it implicitly closes the result set.

Also tweaked something in GameMapper. I was retrieving a non-existent challengerDeck attribute; should've been challenger_deck.
Same for challengee_deck.

Progress for #9

Signed-off-by: vartanbeno <vartanbeno@gmail.com>