Card object has deck_id as parameter, not the whole deck
So basically, the Deck domain object now takes in a List of Cards as a parameter.
If we let the Card object take in a Deck, it makes a StackOverflowError:
- We find the deck
- We find its cards
- For every card, we find its deck
- We find the deck's cards, again
- For every card, we find its deck
- etc. Until we're met with an error.

Adjusted the insert method in the DeckMapper to also insert the list of cards it comes with.
It calls the CardMapper for each card and inserts the card.

Also adjusted the delete method in the DeckMapper to delete its cards from the cards table.

Adjusted /UploadDeck and /ViewDeck to use these domain objects.

See #9

Signed-off-by: vartanbeno <vartanbeno@gmail.com>