Make User object extend SoenEA classes
IUser and User extend IDomainObject<IDField> and DomainObject<IDField>, respectively.
IDField represents the type that we use for the id field, i.e. the primary key of an entity.
It's useful in that it lets us set any type we want. We usually set it to Long, as it is in this case.
Version attribute is of type long also. I had set it to int but I changed it.

UserProxy extends DomainObjectProxy, which acts as a proxy to domain objects (i.e. User).
It has an innerObject variable, storing an instance of the domain object type (i.e. User).
If the innerObject hasn't been initialized yet, it uses the UserMapper to set it to a User object.

UserMapper extends GenericOutputMapper, which provides insert, update, and delete methods to be implemented.

See #16

Signed-off-by: vartanbeno <vartanbeno@gmail.com>