Register servlet -> RegisterDispatcher & RegisterCommand
Every servlet will be replaced into its appropriate dispatcher and command.

The FrontController, which is a sort of parent servlet, will take care of the delegation to the appropriate dispatcher, depending on the path and type of request (GET, POST).

The dispatcher's role is to execute the command. It will then forward the user to the appropriate JSP.
- If everything went well, the user will surely go to a success page.
- If the command threw an exception, the user will go to a fail page.

Also created another createNew() method in the UserFactory, which just takes a username and a password as parameters.

See #21

Signed-off-by: vartanbeno <vartanbeno@gmail.com>