Refactor a ton of code
AbstractCommand (the parent command) now has a @SetInRequestAttribute message variable.
Whenever we want to set the message in the request, we assign it to that variable.

By setting the message in the command, we don't have to set it in the dispatcher.
We just take care of forwarding the user to either success/fail. Message will already be set.

See #21

Signed-off-by: vartanbeno <vartanbeno@gmail.com>