Using /Poke/Deck/\\d+ now to view a deck
FrontController now contains all new endpoints we will use.
A good number of them have parameters in the middle of the URL.

We make sure the path we request to matches the pattern of one that's defined in FrontController, and we return the appropriate dispatcher.
We set the path as a request attribute, so we can access it in the command, from where we extract the desired parameter.

See #30

Signed-off-by: vartanbeno <vartanbeno@gmail.com>