Implement EndTurn logic
I throw an exception if it's not your turn and you attempt to either end your turn, or play a card.

When you end a turn, your opponent immediately draws a card.
The game version is also incremented by 1.

Included version number in the board view of a game.

Also deleted DrawCard dispatcher/command, since we implement that logic upon ending a turn.

See #34

Signed-off-by: vartanbeno <vartanbeno@gmail.com>