Implement playing energy cards
Changed the pokemon_card attribute in for energy cards to just contain
the ID of the benched card, instead of the whole object. The latter was
giving a StackOverflowError when building the AttachedEnergy object,
since Bench objects have a list of AttachedEnergy objects as an
attribute.

You cannot play more than one energy card per turn. We check this by
checking if there are any AttachedEnergy records matching the current
game version (and player, of course).

See #35

Signed-off-by: Vartan Benohanian <vartanbeno@gmail.com>