Upload deck with evolving Pokemon
Added a "basic" attribute in the Card table (only cards of type Pokemon
can have this attribute be set). The card's basic value is the name of
the Pokemon when it's unevolved.

So a Charizard card can have the name Charizard, and the basic value of
Charmeleon.

Also added an example deck with evolving Pokemon in WebContent/

See #41

Signed-off-by: Vartan Benohanian <vartanbeno@gmail.com>