Create predecessor attribute in Bench
It represents the ID of the card that the Pokemon replaces on the bench.
If the card didn't replace another, then we set its predecessor as
itself.

In the game board, if the card has effectively replaced another (i.e. we
made a Pokemon evolve), then we see the ID of the one it replaced,
represented by the "b" attribute in the JSON (stands for 'basic').

We don't see it if the card didn't replace another.

See #44

Signed-off-by: Vartan Benohanian <vartanbeno@gmail.com>