Change game_version attribute in AttachedEnergy to game_turn
Before, to see if we can play an energy card (we can only play 1 per
turn), we compared game versions.

But game versions are incremented not just at the end of a turn, but
whenever a player plays a card. So I created a turn attribute in Game
that ONLY gets incremented once a turn ends. And I use that now to see
if we can play an energy card in a given turn.

See #43

Signed-off-by: Vartan Benohanian <vartanbeno@gmail.com>