Env variables in Tomcat container to set up DB connection
In the `initDB()` method in the FrontController, we try to see if those
environment variables are set (DB_HOST, DB_NAME, etc.).

If one of them is not set, then there's a good chance that the
application isn't being run with `docker-compose`, so we use the
`defaultInitialization()` method, which gets the credentials for the DB
from the `MyResources.properties` file.

This way, we don't have to touch that file ever again when generating
the WAR. Before doing that, I was always changing the hostname to `db`,
to make sure it works in the container. No need to do that now.

Signed-off-by: Vartan Benohanian <vartanbeno@gmail.com>