Remove bench size limit, tweak logic for playing trainer
Bench size limit before this commit was 5. There is now no limit.

If you specify a Pokemon to play a trainer card on, you get a fail
message. Trainer cards just go in the discard pile.

Also made a `docker-compose.db.yml` file, which spawns just 1 container
for the database. No Tomcat container in it.

Closes #55.
Closes #56.

Signed-off-by: Vartan Benohanian <vartanbeno@gmail.com>