Require version number as parameter
For accepting/refusing/withdrawing from a challenge, we require the
version of the challenge, and once we register the challenge as dirty,
the UoW commit checks if the challenge with that ID has the same
version. If it doesn't, then we get a LostUpdateException.

Same with playing a card in a game. Always check the version.

See #32

Signed-off-by: Vartan Benohanian <vartanbeno@gmail.com>