Add health check for MySQL db container
Tomcat container only starts running once the database is up.

Not only does the Tomcat container `depends_on` the `db`, but it waits
for it to be `healthy` to start.

For the `db` to be considered healthy, it tries to connect to the MySQL
db via command line once every 15 seconds, 6 times. As soon as it does
it successfully (i.e. the database can now accept connections), the
Tomcat container starts.

Signed-off-by: Vartan Benohanian <vartanbeno@gmail.com>